pyproject.toml
[project]
name = "storyos"
version = "0.1.0"
description = "Local-first, auditable storytelling workflow engine with secure tool use and plugins."
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1",
]

[project.scripts]
storyos = "storyos.cli:app"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
